Mouse click action for the zones
