+++ web/src/main.js
@@ -35,7 +35,6 @@ import thirdPartyServices from './services/thirdparty';
 
 // Utils
 import { getCurrentZoneData } from './helpers/redux';
-import { centerOnZoneName } from './helpers/interaction';
 
 // Layout
 import Main from './layout/main';
@@ -552,7 +551,7 @@ function renderMap(state) {
     if (selectedZoneName) {
       console.log(`Centering on selectedZoneName ${selectedZoneName}`);
       // eslint-disable-next-line no-use-before-define
-      centerOnZoneName(state, zoneMap, selectedZoneName, 4);
+      centerOnZoneName(state, selectedZoneName, 4);
       hasCenteredMap = true;
     } else if (callerLocation) {
       console.log('Centering on browser location @', callerLocation);
@@ -935,10 +934,6 @@ d3.selectAll('.left-panel-back-button')
 zoneList.setClickHandler((selectedCountry) => {
   dispatchApplication('showPageState', 'country');
   dispatchApplication('selectedZoneName', selectedCountry.countryCode);
-  if (zoneMap !== 'undefined') {
-    // eslint-disable-next-line no-use-before-define
-    centerOnZoneName(getState(), zoneMap, selectedCountry.countryCode, 4);
-  }
 });
 
 // Keyboard navigation
@@ -1275,6 +1270,35 @@ function tryFetchHistory(state) {
   }
 }
 
+function centerOnZoneName(state, zoneName, zoomLevel) {
+  if (typeof zoneMap === 'undefined') { return; }
+  const selectedZone = state.data.grid.zones[zoneName];
+  const selectedZoneCoordinates = [];
+  selectedZone.geometry.coordinates.forEach((geojson) => {
+    // selectedZoneCoordinates.push(geojson[0]);
+    geojson[0].forEach((coord) => {
+      selectedZoneCoordinates.push(coord);
+    });
+  });
+  const maxLon = d3.max(selectedZoneCoordinates, d => d[0]);
+  const minLon = d3.min(selectedZoneCoordinates, d => d[0]);
+  const maxLat = d3.max(selectedZoneCoordinates, d => d[1]);
+  const minLat = d3.min(selectedZoneCoordinates, d => d[1]);
+  const lon = d3.mean([minLon, maxLon]);
+  const lat = d3.mean([minLat, maxLat]);
+
+  zoneMap.setCenter([lon, lat]);
+  if (zoomLevel) {
+    // Remember to set center and zoom in case the map wasn't loaded yet
+    zoneMap.setZoom(zoomLevel);
+    // If the panel is open the zoom doesn't appear perfectly centered because
+    // it centers on the whole window and not just the visible map part.
+    // something one could fix in the future. It's tricky because one has to project, unproject
+    // and project again taking both starting and ending zoomlevel into account
+    zoneMap.map.easeTo({ center: [lon, lat], zoom: zoomLevel });
+  }
+}
+
 function renderExchanges(state) {
   const { exchanges } = state.data.grid;
   const { electricityMixMode } = state.application;
@@ -1447,6 +1471,12 @@ observe(state => state.application.windEnabled, (windEnabled, state) => {
     }
   } else if (typeof windLayer !== 'undefined') {
     windLayer.hide();
+  }
+});
+
+observe(state => state.application.selectedZoneName, (selectedZoneName, state) => {
+  if (selectedZoneName) {
+    centerOnZoneName(state, selectedZoneName, 4);
   }
 });
